added plugin explorer and disabled paste and drag handlers
We figured out to get one panel to react to changes in another panel.
This specific example enables dynamic rendering of plugin markup.

It was challenging to learn the ObservableHQ trick of defining a
variable for 'viewof item' along with the 'item'. The 'item' is
defined with Generators.input() which becomes a way of forwarding the
CustomEvent('input') events to a subscriber.